+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ### 1.2.10
 * allowed input of named colors
 * ignoring data series by clicking table header
+* made MyCharts page searchable
 
 ### 1.2.9
 * added support for S3 domain alias (e.g. s3.datawrapper.de)